Compile fix for X11.
git-svn-id: file:///home/michael/svn/yuni/trunk@1212 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/gfx/surface/cairo.cpp
@@ -17,7 +17,7 @@ namespace Surface
 	void Cairo::release()
 	{}
 
-	void Cairo::resize(unsigned int width, unsigned int height)
+	void Cairo::resize(size_t width, size_t height)
 	{}
 
 	void Cairo::clearColor(Color::RGB<uint8>& newColor)
